Strong name, IWritableFeed in addition to IEnumerableFeed
git-svn-id: https://timeseriesdb.googlecode.com/svn/trunk@89 537f4964-0365-11de-bcc6-d9108be23e13

+++ FastBinTimeseries/Serializers/BlockSerializer/CodecReader.cs
@@ -52,7 +52,7 @@ namespace NYurik.FastBinTimeseries.Serializers.BlockSerializer
         [UsedImplicitly]
         internal int ReadHeader()
         {
-            return ValidateCount(ReadUnsignedValue());
+            return ValidateCount((long) ReadUnsignedValue());
         }
 
         public void AttachBuffer(ArraySegment<byte> value)
